user registration and authentication is implemented

+++ Backend/base/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render,redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login , logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 
 from .models import Room ,Topic
@@ -9,9 +12,12 @@ from . forms import RoomForm
 
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
     
         print(request.POST.get('password'))
@@ -25,11 +31,26 @@ def loginPage(request):
             messages.info(request,'Crediental is invalid')
             return redirect('login')
     else:
-        return render(request,'base/login_register.html')
+        context = {'page':page}
+        return render(request,'base/login_register.html',context)
     
 def logoutUser(request):
     logout(request)
     return redirect('login')
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'an error occured during registration')
+    return render(request,'base/login_register.html',{'form':form})
 
 
 def home(request):
@@ -56,7 +77,7 @@ def room(request,pk):
              
     return render(request,'base/room.html',context )
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
      
@@ -69,9 +90,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'base/room_form.html',context) 
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room  = Room.objects.get(id = pk)
     form  = RoomForm(instance = room)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!.')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
         if form.is_valid:
@@ -82,8 +107,12 @@ def updateRoom(request,pk):
     
     return render(request,'base/room_form.html',context )
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id =pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!.')
+    
     if request.method == "POST":
          room.delete()
          return redirect('home')
